+++ src/starkbank/payment_request.clj
@@ -7,7 +7,7 @@
 
   ## Parameters (required):
   - `:center-id` [string]: unique id returned when PaymentRequest is created. ex: \"5656565656565656\"
-  - `:payment` [Transfer, BrcodePayment, BoletoPayment, TaxPayment, UtilityPayment, Transaction or map]: payment entity that should be approved and executed.
+  - `:payment` [Transfer, BrcodePayment, BoletoPayment, DarfPayment, TaxPayment, UtilityPayment, Transaction or map]: payment entity that should be approved and executed.
   - `:type` [string]: payment type, inferred from the payment parameter if it is not a map. ex: \"transfer\", \"brcode-payment\"
 
   ## Parameters (optional):
@@ -28,6 +28,7 @@
             [starkbank.transaction]
             [starkbank.brcode-payment]
             [starkbank.boleto-payment]
+            [starkbank.darf-payment]
             [starkbank.tax-payment]
             [starkbank.utility-payment])
   (:use [clojure.walk]))
@@ -55,6 +56,7 @@
         "transaction" (#'starkbank.transaction/java-to-clojure (.payment java-object))
         "brcode-payment" (#'starkbank.brcode-payment/java-to-clojure (.payment java-object))
         "boleto-payment" (#'starkbank.boleto-payment/java-to-clojure (.payment java-object))
+        "darf-payment" (#'starkbank.darf-payment/java-to-clojure (.payment java-object))
         "tax-payment" (#'starkbank.tax-payment/java-to-clojure (.payment java-object))
         "utility-payment" (#'starkbank.utility-payment/java-to-clojure (.payment java-object)))
     }))
@@ -74,6 +76,7 @@
           "transaction" (#'starkbank.transaction/clojure-to-java payment)
           "brcode-payment" (#'starkbank.brcode-payment/clojure-to-java payment)
           "boleto-payment" (#'starkbank.boleto-payment/clojure-to-java payment)
+          "darf-payment" (#'starkbank.darf-payment/clojure-to-java payment)
           "tax-payment" (#'starkbank.tax-payment/clojure-to-java payment)
           "utility-payment" (#'starkbank.utility-payment/clojure-to-java payment))
         "type" type