+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - page functions as a manual-pagination alternative to queries
 - Institution resource to allow query of institutions recognized by the Brazilian Central Bank for Pix and TED transactions
 - TaxPayment resource
+- DarfPayment resource
 
 ## [2.3.4] - 2021-03-04
 ### Added